requirements.txt
asgiref==3.4.1
atomicwrites==1.4.1
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==2.0.12
colorama==0.4.5
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.2.0
Django==3.2.25
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==4.4.0
environ==1.0
filelock==3.4.1
gunicorn==21.2.0
httpie==2.6.0
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
isoweek==1.3.3
jsons==1.6.3
mysqlclient==2.0.0
packaging==21.3
pipenv==2022.4.8
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.8.3
py==1.11.0
Pygments==2.14.0
PyJWT==2.4.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==7.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-toolbelt==1.0.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tomli==1.2.3
typing==3.7.4.3
typing_extensions==4.1.1
typish==1.9.3
tzdata==2024.1
tzlocal==4.2
urllib3==1.26.19
virtualenv==20.17.1
virtualenv-clone==0.5.7
xlwt==1.3.0
zipp==3.6.0
